# swagger.yaml
openapi: 3.0.0
info:
  title: Anime Characters API
  version: 1.0.0
  description: API para listar personagens de animes

servers:
  - url: https://animesapi-production-afranio.up.railway.app
    description: Servidor de produção (Railway)

paths:
  /auth/register:
    post:
      summary: Registra um novo usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nome:
                  type: string
                email:
                  type: string
                senha:
                  type: string
      responses:
        '200':
          description: Usuário registrado com sucesso

  /auth/login:
    post:
      summary: Faz login e retorna um token JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                senha:
                  type: string
      responses:
        '200':
          description: Token JWT
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string

  /api/characters:
    get:
      summary: Lista todos os personagens
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de personagens
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Character'

  /api/characters/{id}:
    get:
      summary: Busca um personagem pelo ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID do personagem
          schema:
            type: integer
      responses:
        '200':
          description: Personagem encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Character'
        '404':
          description: Personagem não encontrado

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Character:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        anime:
          type: string